backend/requirements.txt
# Core
Django==5.0.1
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9

# Authentication
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.1

# CORS
django-cors-headers==4.3.1

# ML & Data Processing
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3

# Real-time Features
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1

# API Documentation
drf-spectacular==0.26.5